+++ Source/Blazorise/Check.razor.cs
@@ -23,6 +23,19 @@ namespace Blazorise
 
         #region Methods
 
+        public override async Task SetParametersAsync( ParameterView parameters )
+        {
+            await base.SetParametersAsync( parameters );
+
+            if ( ParentValidation != null )
+            {
+                if ( parameters.TryGetValue<Expression<Func<TValue>>>( nameof( CheckedExpression ), out var expression ) )
+                    ParentValidation.InitializeInputExpression( expression );
+
+                InitializeValidation();
+            }
+        }
+
         protected override void BuildClasses( ClassBuilder builder )
         {
             builder.Append( ClassProvider.Check() );
